+++ Source/TDGame/Public/TDFoe.h
@@ -119,7 +119,7 @@ protected:
 	FRotator OriginalRotation;
 	//				  ******HANDLE DAMAGE & MAKE DAMAGE******
 	UFUNCTION()
-	void HandleTakeDamage(UTDHealthComponent* OwningHealthComp, float Health, float HealthDelta, const class UDamageType* DamageType,
+	void HandleTakeDamage(UTDHealthComponent* OwningHealthComp, int32 Health, int32 HealthDelta, const class UDamageType* DamageType,
 		class AController* InstigatedBy, AActor* DamageCauser);
 
 	UMaterialInstanceDynamic* MatInst;
